requirements.txt
pyzmq==16.0.4
netaddr==0.7.19
networkx==2.1
blosc==1.5.1